requirements.txt
setuptools==75.3.0
pandas==2.2.3
requests==2.32.3
bs4==0.0.2
beautifulsoup4==4.12.3
newspaper3k==0.2.8
news-please==1.6.13
Unidecode==1.3.8
selenium==4.26.1
twine==5.1.1